0.3.0 Add town-level resourceProductionModifier to resource production.
Towns now have their production amounts altered based on the TownLevel
resourceProductionModifier.

Resources are displayed with multiplier-altered amounts in the
dynmap-towny infowindows and in the town status screen.

The Town status screen now comes with a modifier % when a modification
is made.

When towns earn resources on a new day, the amount given is altered
based on their current town level modifier.

Bump min. Towny version to 0.98.6.3.

Bump lang files to 0.10.